Create default functions and operator info
Binary operations such as addition (+) can apply to multiple types e.g. int + long, double + BigDecimal, etc. Make a BinaryOperatorInfo class that holds the operator symbol (e.g. +), and two java.lang.Class objects for the parameter types. That class implements and generates a hash code on the operator and the two parameters. The evaluator will execute the correct binary function based on the operation of the binary node and types of the two operands.