Make a workaround in Expression for Gradle build
The Expression class contains a val(T o, Class<T> cls) method. There is a switch statement that switches on the cls and all of the cases do a primitive cast e.g. (boolean) o. Gradle build fails because of unsafe operations. Unfortunately, there is no way to disable the linter since compilerOptions is read-only.