Use HashMap of binary nodes by precedence
This helps to insert binary operator nodes of any priority, relative to existing operator nodes depending on their priority, to the correct positions within the tree.

The processDeque field was removed. The reason why it was there in the first place was because it was carried over from an earlier attempt (early 2014) at creating a binary expression tree. That earlier source code is inside a "HelperLib" project which contains utility functions for Java. I will not put it up on GitHub because it is not as flexible or well written as this library. In the earlier code, operators were hard coded, and boolean and math expressions had their own builder classes.