Solve the commutative property issue
For non-commutative operations such as subtraction and division, and for different operand types, the only solution was to create two functions where the only difference was that the types were switched between the left and right operands. For commutative operations, to prevent that duplication, I switched the types to find the correct function.