Implement variables in nested expressions
Add Variable objects from nested Expressions to the containing Expression's Variable HashMap. Throw an error when the nested expression contains a different (i.e. the == reference equality check fails) Variable with the same name as a Variable in the containing Expression.

Create a new Expression for the nested expression when expr() is called because the caller should not be able to mutate the nested expression.

Fix a bug where an error would have been thrown if the Expression object that is passed into flatExpr() method had a nested expression.